# src/docs/swagger.yaml
definitions:
  dto.CityResponse:
    properties:
      country:
        $ref: '#/definitions/dto.CountryResponse'
      id:
        type: integer
      name:
        type: string
    type: object
  dto.ColorResponse:
    properties:
      hex:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  dto.CountryResponse:
    properties:
      name:
        type: string
    type: object
  dto.CreateCityRequest:
    properties:
      countryId:
        type: integer
      name:
        type: string
    required:
    - countryId
    - name
    type: object
  dto.CreateColorRequest:
    properties:
      hex:
        type: string
      name:
        type: string
    type: object
  dto.CreateCountryRequest:
    properties:
      name:
        type: string
    type: object
  dto.CreatePersianYearRequest:
    properties:
      name:
        type: string
      year:
        type: integer
    type: object
...
  dto.PersianYearResponse:
    properties:
      id:
        type: integer
      name:
        type: string
      year:
        type: integer
    type: object
...
  dto.UpdateCityRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  dto.UpdateColorRequest:
    properties:
      hex:
        type: string
      name:
        type: string
    type: object
  dto.UpdateCountryRequest:
    properties:
      name:
        type: string
    type: object
  dto.UpdatePersianYearRequest:
    properties:
      name:
        type: string
    type: object
...
  helper.Response:
    properties:
      error: {}
      result: {}
      result_code:
        type: integer
      success:
        type: boolean
      validation_error:
        items:
          $ref: '#/definitions/validators.ValidationError'
        type: array
    type: object
  validators.ValidationError:
    properties:
      message:
        type: string
      property:
        type: string
      tag:
        type: string
      value:
        type: string
    type: object
info:
  contact: {}
paths:
  /v1/city/create:
    post:
      consumes:
      - application/json
      description: Create a City
      parameters:
      - description: Create a City
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCityRequest'
      responses:
        "201":
          description: City response
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                result:
                  $ref: '#/definitions/dto.CityResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: Create a City
      tags:
      - City
  /v1/city/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Get a City
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: City response
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: Get a City
      tags:
      - City
  /v1/city/get/{id}:
    get:
      consumes:
      - application/json
      description: Delete a City
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: response
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: Delete a City
      tags:
      - City
  /v1/city/update/{id}:
    put:
      consumes:
      - application/json
      description: Update a City
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      - description: Update a City
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCityRequest'
      responses:
        "200":
          description: City response
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                result:
                  $ref: '#/definitions/dto.CityResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: Update a City
      tags:
      - City
  /v1/color/create:
    post:
      consumes:
      - application/json
      description: Create a Color
      parameters:
      - description: Create a Color
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateColorRequest'
      responses:
        "201":
          description: Color response
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                result:
                  $ref: '#/definitions/dto.ColorResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: Create a Color
      tags:
      - Color
  /v1/color/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Get a Color
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Color response
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: Get a Color
      tags:
      - Color
  /v1/color/get/{id}:
    get:
      consumes:
      - application/json
      description: Delete a Color
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: response
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: Delete a Color
      tags:
      - Color
  /v1/color/update/{id}:
    put:
      consumes:
      - application/json
      description: Update a Color
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      - description: Update a Color
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateColorRequest'
      responses:
        "200":
          description: Color response
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                result:
                  $ref: '#/definitions/dto.ColorResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: Update a Color
      tags:
      - Color
  /v1/country/create:
    post:
      consumes:
      - application/json
      description: Create a Country
      parameters:
      - description: Create a Country
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCountryRequest'
      responses:
        "201":
          description: Country response
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                result:
                  $ref: '#/definitions/dto.CountryResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: Create a Country
      tags:
      - Country
  /v1/country/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Get a Country
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Country response
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: Get a Country
      tags:
      - Country
  /v1/country/get/{id}:
    get:
      consumes:
      - application/json
      description: Delete a Country
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: response
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: Delete a Country
      tags:
      - Country
  /v1/country/update/{id}:
    put:
      consumes:
      - application/json
      description: Update a Country
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      - description: Update a Country
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCountryRequest'
      responses:
        "200":
          description: Country response
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                result:
                  $ref: '#/definitions/dto.CountryResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: Update a Country
      tags:
      - Country
...
  /v1/year/create:
    post:
      consumes:
      - application/json
      description: Create a PersianYear
      parameters:
      - description: Create a PersianYear
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePersianYearRequest'
      responses:
        "201":
          description: PersianYear response
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                result:
                  $ref: '#/definitions/dto.PersianYearResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: Create a PersianYear
      tags:
      - PersianYear
  /v1/year/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Get a PersianYear
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: PersianYear response
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: Get a PersianYear
      tags:
      - PersianYear
  /v1/year/get/{id}:
    get:
      consumes:
      - application/json
      description: Delete a PersianYear
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: response
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: Delete a PersianYear
      tags:
      - PersianYear
  /v1/year/update/{id}:
    put:
      consumes:
      - application/json
      description: Update a PersianYear
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      - description: Update a PersianYear
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePersianYearRequest'
      responses:
        "200":
          description: PersianYear response
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                result:
                  $ref: '#/definitions/dto.PersianYearResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: Update a PersianYear
      tags:
      - PersianYear
swagger: "2.0"